migrated signup to user route js
