+++ app.js
@@ -38,7 +38,6 @@ store.sync();
 
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
-app.use('/signUp', signUpRouter);
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
